requirements.txt
discord.py >= 1.6.0
aiosqlite >= 0.17.0
discord-py-slash-command >= 1.2.0
